# pkg/replace: make it more efficient.