@miekg 2019-02-09
Looks like replacer needs another test to increase coverage

On Sat, 9 Feb 2019, 11:49 Codecov <notifications@github.com wrote:

> Codecov <https://codecov.io/gh/coredns/coredns/pull/2544?src=pr&el=h1>
> Report
>
> Merging #2544
> <https://codecov.io/gh/coredns/coredns/pull/2544?src=pr&el=desc> into
> master
> <https://codecov.io/gh/coredns/coredns/commit/ab4262d7ebaf53639feb0540db40b9298cf3c38e?src=pr&el=desc>
> will *decrease* coverage by 0.35%.
> The diff coverage is 44.73%.
>
> [image: Impacted file tree graph]
> <https://codecov.io/gh/coredns/coredns/pull/2544?src=pr&el=tree>
>
> @@            Coverage Diff             @@
> ##           master    #2544      +/-   ##
>
> ==========================================
> - Coverage    55.4%   55.05%   -0.36%
>
> ==========================================
>
>   Files         204      204
>
>   Lines       10349    10357       +8
>
> ==========================================
> - Hits         5734     5702      -32
> - Misses       4196     4239      +43
> + Partials      419      416       -3
>
> Impacted Files
> <https://codecov.io/gh/coredns/coredns/pull/2544?src=pr&el=tree> Coverage
> Δ
> plugin/rewrite/condition.go
> <https://codecov.io/gh/coredns/coredns/pull/2544/diff?src=pr&el=tree#diff-cGx1Z2luL3Jld3JpdGUvY29uZGl0aW9uLmdv> 0%
> <0%> (ø) ⬆️
> plugin/log/setup.go
> <https://codecov.io/gh/coredns/coredns/pull/2544/diff?src=pr&el=tree#diff-cGx1Z2luL2xvZy9zZXR1cC5nbw==> 88.05%
> <0%> (ø) ⬆️
> plugin/log/log.go
> <https://codecov.io/gh/coredns/coredns/pull/2544/diff?src=pr&el=tree#diff-cGx1Z2luL2xvZy9sb2cuZ28=> 88.46%
> <100%> (ø) ⬆️
> plugin/pkg/replacer/replacer.go
> <https://codecov.io/gh/coredns/coredns/pull/2544/diff?src=pr&el=tree#diff-cGx1Z2luL3BrZy9yZXBsYWNlci9yZXBsYWNlci5nbw==> 29.62%
> <46.37%> (-32.38%) ⬇️
> plugin/errors/errors.go
> <https://codecov.io/gh/coredns/coredns/pull/2544/diff?src=pr&el=tree#diff-cGx1Z2luL2Vycm9ycy9lcnJvcnMuZ28=> 95.55%
> <0%> (-4.45%) ⬇️
> ------------------------------
>
> Continue to review full report at Codecov
> <https://codecov.io/gh/coredns/coredns/pull/2544?src=pr&el=continue>.
>
> *Legend* - Click here to learn more
> <https://docs.codecov.io/docs/codecov-delta>
> Δ = absolute <relative> (impact), ø = not affected, ? = missing data
> Powered by Codecov
> <https://codecov.io/gh/coredns/coredns/pull/2544?src=pr&el=footer>. Last
> update ab4262d...c324e2c
> <https://codecov.io/gh/coredns/coredns/pull/2544?src=pr&el=lastupdated>.
> Read the comment docs <https://docs.codecov.io/docs/pull-request-comments>
> .
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/coredns/coredns/pull/2544#issuecomment-462037796>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAVkW2PL3ZS52k9G9LDE0JCyFFwzj-Nwks5vLrXJgaJpZM4aygUj>
> .
>